requirements.txt
flask>=3.0,<4.0
langchain>=1.2,<2.0
langchain-text-splitters>=1.0,<2.0
chromadb>=0.4.0,<1.0.0
sentence-transformers>=5.0,<6.0
pypdf>=6.0,<7.0
python-docx>=1.1,<2.0
python-dotenv>=1.0,<2.0
groq>=0.9.0
gunicorn>=21.0,<26.0
